Feat: added css to header
